# Conflicts:
#	.pre-commit-config.yaml
#	src/engine/controls/bpmcontrol.h
#	src/mixer/basetrackplayer.cpp
#	src/test/beatgridtest.cpp
#	src/test/beatmaptest.cpp
#	src/test/enginesynctest.cpp
#	src/track/beatgrid.cpp
#	src/track/beatgrid.h
#	src/track/beatmap.cpp
#	src/track/beatmap.h
#	src/track/track.cpp
#	src/track/track.h
#	src/widget/wtracktableview.cpp

+++ src/mixxx.cpp
@@ -202,6 +202,7 @@ MixxxMainWindow::MixxxMainWindow(QApplication* pApp, const CmdlineArgs& args)
 }
 
 MixxxMainWindow::~MixxxMainWindow() {
+    finalize();
     // SkinLoader depends on Config;
     delete m_pSkinLoader;
 }
@@ -305,6 +306,10 @@ void MixxxMainWindow::initialize(QApplication* pApp, const CmdlineArgs& args) {
             &PlayerManager::noMicrophoneInputConfigured,
             this,
             &MixxxMainWindow::slotNoMicrophoneInputConfigured);
+    connect(m_pPlayerManager,
+            &PlayerManager::noAuxiliaryInputConfigured,
+            this,
+            &MixxxMainWindow::slotNoAuxiliaryInputConfigured);
     connect(m_pPlayerManager,
             &PlayerManager::noDeckPassthroughInputConfigured,
             this,
@@ -814,7 +819,10 @@ void MixxxMainWindow::finalize() {
     t.elapsed(true);
     // Report the total time we have been running.
     m_runtime_timer.elapsed(true);
-    StatsManager::destroy();
+
+    if (m_cmdLineArgs.getDeveloper()) {
+        StatsManager::destroy();
+    }
 }
 
 bool MixxxMainWindow::initializeDatabase() {
@@ -1354,11 +1362,24 @@ void MixxxMainWindow::slotNoDeckPassthroughInputConfigured() {
 }
 
 void MixxxMainWindow::slotNoMicrophoneInputConfigured() {
-    QMessageBox::StandardButton btn = QMessageBox::warning(
+    QMessageBox::StandardButton btn = QMessageBox::question(
         this,
         Version::applicationName(),
         tr("There is no input device selected for this microphone.\n"
-           "Please select an input device in the sound hardware preferences first."),
+           "Do you want to select an input device?"),
+        QMessageBox::Ok | QMessageBox::Cancel, QMessageBox::Cancel);
+    if (btn == QMessageBox::Ok) {
+        m_pPrefDlg->show();
+        m_pPrefDlg->showSoundHardwarePage();
+    }
+}
+
+void MixxxMainWindow::slotNoAuxiliaryInputConfigured() {
+    QMessageBox::StandardButton btn = QMessageBox::question(
+        this,
+        Version::applicationName(),
+        tr("There is no input device selected for this auxiliary.\n"
+           "Do you want to select an input device?"),
         QMessageBox::Ok | QMessageBox::Cancel, QMessageBox::Cancel);
     if (btn == QMessageBox::Ok) {
         m_pPrefDlg->show();
@@ -1511,7 +1532,6 @@ void MixxxMainWindow::closeEvent(QCloseEvent *event) {
         event->ignore();
         return;
     }
-    finalize();
     QMainWindow::closeEvent(event);
 }
